updated user controller

+++ backend/controllers/user.controller.js
@@ -1,14 +1,80 @@
-import bcrypt from 'bcryptjs';
 import User from '../models/user.model.js';
 import Follow from '../models/follow.model.js';
+import bcrypt from 'bcryptjs';
 import jwt from 'jsonwebtoken';
+
+export const registerUser = async (req, res) => {
+  const { username, displayName, email, password } = req.body;
+
+  if (!username || !email || !password) {
+    return res.status(400).json({ message: 'All fields are required!' });
+  }
+
+  const newHashedPassword = await bcrypt.hash(password, 10);
+
+  const user = await User.create({
+    username,
+    displayName,
+    email,
+    hashedPassword: newHashedPassword,
+  });
+
+  const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET_KEY);
+
+  res.cookie('token', token, {
+    httpOnly: true,
+    secure: process.env.NODE_ENV === 'production',
+    maxAge: 30 * 24 * 60 * 60 * 1000,
+  });
+
+  const { hashedPassword, ...detailsWithoutPassword } = user.toObject();
+
+  res.status(201).json(detailsWithoutPassword);
+};
+export const loginUser = async (req, res) => {
+  const { email, password } = req.body;
+
+  if (!email || !password) {
+    return res.status(400).json({ message: 'All fields are required!' });
+  }
+
+  const user = await User.findOne({ email });
+
+  if (!user) {
+    return res.status(401).json({ message: 'Invalid email or password' });
+  }
+
+  const isPasswordCorrect = await bcrypt.compare(password, user.hashedPassword);
+
+  if (!isPasswordCorrect) {
+    return res.status(401).json({ message: 'Invalid email or password' });
+  }
+
+  const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET_KEY);
+
+  res.cookie('token', token, {
+    httpOnly: true,
+    secure: process.env.NODE_ENV === 'production',
+    maxAge: 30 * 24 * 60 * 60 * 1000,
+  });
+
+  const { hashedPassword, ...detailsWithoutPassword } = user.toObject();
+
+  res.status(200).json(detailsWithoutPassword);
+};
+
+export const logoutUser = async (req, res) => {
+  res.clearCookie('token');
+
+  res.status(200).json({ message: 'Logout successful' });
+};
 
 export const getUser = async (req, res) => {
   const { username } = req.params;
 
   const user = await User.findOne({ username });
 
-  const { hashedPassword, ...detailsWithPassword } = user.toObject();
+  const { hashedPassword, ...detailsWithoutPassword } = user.toObject();
 
   const followerCount = await Follow.countDocuments({ following: user._id });
   const followingCount = await Follow.countDocuments({ follower: user._id });
@@ -17,7 +83,7 @@ export const getUser = async (req, res) => {
 
   if (!token) {
     res.status(200).json({
-      ...detailsWithPassword,
+      ...detailsWithoutPassword,
       followerCount,
       followingCount,
       isFollowing: false,
@@ -31,7 +97,7 @@ export const getUser = async (req, res) => {
         });
 
         res.status(200).json({
-          ...detailsWithPassword,
+          ...detailsWithoutPassword,
           followerCount,
           followingCount,
           isFollowing: isExists ? true : false,
@@ -52,81 +118,10 @@ export const followUser = async (req, res) => {
   });
 
   if (isFollowing) {
-    await Follow.deleteOne({
-      follower: req.userId,
-      following: user._id,
-    });
+    await Follow.deleteOne({ follower: req.userId, following: user._id });
   } else {
-    await Follow.create({
-      follower: req.userId,
-      following: user._id,
-    });
+    await Follow.create({ follower: req.userId, following: user._id });
   }
 
-  res.status(200).json({ message: 'success' });
-};
-
-export const registerUser = async (req, res) => {
-  const { username, displayName, email, password } = req.body;
-  if (!username || !email || !password || !displayName) {
-    return res.status(400).json({ message: 'Please enter all fields' });
-  }
-
-  const newHashedPassword = await bcrypt.hash(password, 10);
-
-  const user = await User.create({
-    username,
-    displayName,
-    email,
-    hashedPassword: newHashedPassword,
-  });
-
-  const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET_KEY);
-
-  res.cookie('token', token, {
-    httpOnly: true,
-    secure: process.env.NODE_ENV === 'production',
-    maxAge: 30 * 24 * 60 * 60 * 1000,
-  });
-
-  const { hashedPassword, ...detailsWithPassword } = user.toObject();
-
-  res.status(201).json(detailsWithPassword);
-};
-
-export const loginUser = async (req, res) => {
-  const { email, password } = req.body;
-  if (!email || !password) {
-    return res.status(400).json({ message: 'Please enter all fields' });
-  }
-
-  const user = await User.findOne({ email });
-
-  if (!user) {
-    return res.status(400).json({ message: 'Invalid email or password' });
-  }
-
-  const isPasswordSame = await bcrypt.compare(password, user.hashedPassword);
-
-  if (!isPasswordSame) {
-    return res.status(400).json({ message: 'Invalid email or password' });
-  }
-
-  const token = jwt.sign({ userId: user._id }, process.env.JWT_SECRET_KEY);
-
-  res.cookie('token', token, {
-    httpOnly: true,
-    secure: process.env.NODE_ENV === 'production', // Secure in production
-    sameSite: 'None', // Needed for cross-site cookies
-    maxAge: 30 * 24 * 60 * 60 * 1000,
-  });
-
-  const { hashedPassword, ...detailsWithPassword } = user.toObject();
-
-  res.status(201).json(detailsWithPassword);
-};
-
-export const logoutUser = async (req, res) => {
-  res.clearCookie('token');
-  return res.status(200).json({ message: 'Logout successfully' });
+  res.status(200).json({ message: 'Successful' });
 };
